examples/python/oracle_arb_finder/requirements.txt
web3==5.13.1
python-dotenv==0.15.0
